refactor: Problem 목록조회 정적 팩토리 메소드 사용하도록 변경
Related to : #9
