+++ backend/src/main/java/com/prograngers/backend/dto/problem/ProblemResponse.java
@@ -1,9 +1,7 @@
-package com.prograngers.backend.repository.problem.dto;
+package com.prograngers.backend.dto.problem;
 
 import com.prograngers.backend.entity.Problem;
 import com.prograngers.backend.entity.Solution;
-import com.prograngers.backend.entity.constants.AlgorithmConstant;
-import com.prograngers.backend.entity.constants.DataStructureConstant;
 import com.prograngers.backend.entity.constants.JudgeConstant;
 import lombok.Builder;
 import lombok.Getter;
