+++ backend/src/main/java/com/prograngers/backend/repository/problem/QueryDslProblemRepositoryImpl.java
@@ -1,18 +1,15 @@
 package com.prograngers.backend.repository.problem;
 
 import com.prograngers.backend.entity.Problem;
-import com.prograngers.backend.entity.Solution;
 import com.prograngers.backend.entity.constants.AlgorithmConstant;
 import com.prograngers.backend.entity.constants.DataStructureConstant;
-import com.prograngers.backend.repository.problem.dto.ProblemResponse;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
-import java.util.ArrayList;
-import java.util.HashMap;
+
 import java.util.List;
 
 
@@ -26,7 +23,7 @@ public class QueryDslProblemRepositoryImpl implements QueryDslProblemRepository 
 
     private final JPAQueryFactory jpaQueryFactory;
 
-    public List<ProblemResponse> searchByAlgorithmAndDataStructureOrderByDateDesc(
+    public List<Problem> searchByAlgorithmAndDataStructureOrderByDateDesc(
             int page, DataStructureConstant dataStructure, AlgorithmConstant algorithm, String orderBy) {
         // 양방향 연관관계로 변경
         List<Problem> results = jpaQueryFactory
@@ -35,16 +32,16 @@ public class QueryDslProblemRepositoryImpl implements QueryDslProblemRepository 
                 .join(problem.solutions, solution).fetchJoin()
                 .where(dataStructureEq(dataStructure), algorithmEq(algorithm))
                 .orderBy(orderByWhat(orderBy))
-                .offset((page-1)*4)
+                .offset((page - 1) * 4)
                 .limit(4)
                 .fetch();
-        return  ProblemResponse.toDto(results);
+        return results;
     }
 
     private OrderSpecifier<?> orderByWhat(String orderBy) {
-        if (orderBy.equals("date")){ // date 인 경우
+        if (orderBy.equals("date")) { // date 인 경우
             log.info("orderByDate");
-           return problem.date.desc();
+            return problem.date.desc();
             // size-1의 solution (제일 마지막 solution)의 날짜  기준으로 정렬
 //            NumberExpression<Integer> size = problem.solutions.size().subtract(1);
 //           return problem.solutions.get(size).date.desc();
